moved shp.hpp one level up fixed fields initialization order in dense_matrix_view
