+++ examples/shp/CMakeLists.txt
@@ -16,6 +16,14 @@ endfunction()
 
 add_shp_example(vector_example)
 add_shp_example(test_range)
+add_shp_example(copy_test)
 add_shp_example(dot_product)
 add_shp_example(dot_product_benchmark)
 add_shp_example(inclusive_scan_example)
+# unsatisfied dependency of grb/grb.hpp
+# add_shp_example(gemv_benchmark)
+add_shp_example(gemv_example)
+add_shp_example(matrix_example)
+add_shp_example(sparse_test)
+add_shp_example(take_example)
+add_shp_example(zip_example)
