+++ examples/shp/dot_product_benchmark.cpp
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: BSD-3-Clause
 
-#include <dr/shp/shp.hpp>
+#include <dr/shp.hpp>
 #include <fmt/core.h>
 #include <fmt/ranges.h>
 #include <oneapi/dpl/algorithm>
