+++ examples/shp/sparse_test.cpp
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: BSD-3-Clause
 
 #include <cstdio>
-#include <dr/shp/shp.hpp>
+#include <dr/shp.hpp>
 
 int main(int argc, char **argv) {
   auto devices = dr::shp::get_numa_devices(sycl::default_selector_v);
