+++ include/dr/shp.hpp
@@ -4,10 +4,6 @@
 
 #pragma once
 
-#include <dr/detail/iterator_adaptor.hpp>
-#include <dr/detail/normal_distributed_iterator.hpp>
-#include <dr/detail/ranges_shim.hpp>
-#include <dr/detail/segments_tools.hpp>
 #include <dr/shp/algorithms/algorithms.hpp>
 #include <dr/shp/containers/dense_matrix.hpp>
 #include <dr/shp/containers/sparse_matrix.hpp>
@@ -17,7 +13,4 @@
 #include <dr/shp/init.hpp>
 #include <dr/shp/range.hpp>
 #include <dr/shp/range_adaptors.hpp>
-#include <dr/shp/util.hpp>
-#include <dr/shp/util/matrix_io.hpp>
 #include <dr/shp/views/views.hpp>
-#include <dr/views/views.hpp>
